reconfigure component and modified files
